backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
pydantic>=2.5.0,<2.7.0
pydantic-settings==2.1.0
redis==5.0.1
python-multipart==0.0.7
httpx==0.26.0
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiogram>=3.4.0
openai>=1.10.0
gigachat>=0.1.0

SpeechRecognition

pydub
rq==1.16.1

# Testing
aiosqlite>=0.19.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
testcontainers>=4.0
websockets>=12.0

# Security
python-dotenv>=1.0.0
slowapi>=0.1.9

# Observability
prometheus-client>=0.20.0
sentry-sdk[fastapi]>=1.39.0
sentry-sdk[fastapi]>=1.40.0

# Scheduler
apscheduler>=3.10.0
